HW - Integración de API de LLM/requirements.txt
httpx>=0.25.0
trafilatura>=1.6.0
rich>=13.7.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
respx>=0.20.0
openai>=1.40.0
